requirements.txt
gensim
numpy
nltk
flask
bs4
requests
pillow
